requirements.txt
# Core dependencies
dspy-ai>=2.6.8
openai>=1.65.0
python-dotenv>=1.0.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0

# Deep learning and transformers
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Vector stores and retrieval
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Evaluation and metrics
scikit-learn>=1.3.0
nltk>=3.8.0